Moved /kick into own file
