+++ src/shout_subgroup/group_chat_listener.py
@@ -16,7 +16,7 @@ from shout_subgroup.repository import (
     insert_group_chat
 )
 from shout_subgroup.utils import is_group_chat
-from shout_subgroup.modify_subgroup import remove_users_from_existing_subgroup
+from shout_subgroup.remove_subgroup_members import remove_users_from_existing_subgroup
 
 
 async def add_user_to_group_chat(db: Session, chat: Chat, current_user: UserModel) -> UserModel | None:
