+++ src/shout_subgroup/main.py
@@ -5,7 +5,8 @@ from telegram.ext import ApplicationBuilder, CommandHandler, filters, MessageHan
 
 from group_chat_listener import listen_for_messages_handler, listen_for_new_member_handler, \
     listen_for_left_member_handler
-from modify_subgroup import subgroup_handler, remove_subgroup_member_handler
+from modify_subgroup import subgroup_handler
+from shout_subgroup.remove_subgroup_members import remove_subgroup_member_handler
 from shout import shout_handler
 from shout_subgroup.delete_subgroup import remove_subgroup_handler
 from shout_subgroup.list_subgroup import list_subgroup_handler
